requirements.txt
discord.py
firebase-admin
python-dotenv
requests